+++ packages/serinus/CHANGELOG.md
@@ -14,3 +14,9 @@ Initial Implementation
 - Added more exceptions
 - Added ApplicationInit mixins
 - Refactoring of the code and small fixes
+
+## 0.0.1-dev.4
+
+- Huge refactoring of the code
+- Add guards for the request
+- Add pipes for the request